Removed Large comment and added gryo code
I removed a comment that had the functions of commWithArduino. I had the
comment in the first place just in case if function.py would fail.
However, it is working fantastically so yeah. Also I added code that can
manipulate the sensativity of the gryo sensor. Making it easier to test
and experiment