Gryo Orientation System Change
Since the gyro readings are now functionally, we have began making a
system to account the -180 - 180 border in the gyro spectrum and also
added limits to how much the robot can attemp to fix while moving. If
the robot is off by 45 degrees then it stops moving and fix its
displacement stationary.